+++ src/main/scala/ai/tripl/arc/config/package.scala
@@ -2,8 +2,14 @@ package ai.tripl.arc
 
 package object config {
 
-  type StringList = List[String]
+  type DoubleList = List[Double]
+
+  type BooleanList = List[Boolean]
 
   type IntList = List[Int]
+
+  type LongList = List[Long]
+
+  type StringList = List[String]
 
 }